requirements.txt
aniso8601==9.0.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
firebase-admin==6.2.0
Flask==2.3.2
Flask-RESTful==0.3.10
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.3
msgpack==1.0.5
oauth2client==3.0.0
peewee==3.16.2
proto-plus==1.22.3
protobuf==4.24.0rc1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.4.3
PyJWT==2.7.0
pyparsing==3.1.0
Pyrebase==3.0.18
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sseclient==0.0.27
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
